pyproject.toml
[tool.poetry]
name = "task-1-final"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
psycopg2 = "^2.9.9"
alembic = "^1.13.2"
python-dotenv = "^1.0.1"
pylint = "^3.2.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
